some housekeeping / restructuring
